A working, though not pretty, scatterplot with new data
After some data wrangling I was able to get the cereal.csv example code
working with a new, more journalistically-focused dataset: one that
contrasts per-pupil expenditure with MCAS scores.

The previous datasets had problems -- one column (total expenditure)
was showing up as NaN (not a number), and after examining the
.CSV, the numbers in that column were enclosed in double quotes.

Removing the quotes got the numbers to show up, but the numbers were
notated as 14,457, which meant that the scatterplot showed the expenditure
as 14, not $14,457. Fixed that too, and then got a plot.

Next steps:

Y axis shows values with small expenditures on top and large ones on the bottom.
It needs to be reversed.

Legend is not useful and should be removed. Puts list of
cities on the left, unreadable. Use tooltips instead.