Working scatterplot, but need to change domain/range
Scatterplot works now, but isn't very readable/useful.

Most of the plot is empty, with all dots clustered closely on the left.

Need to have X axis start at 400 and go to 550; need to reread items on
domain/range to figure out best way to do that.

Also need to clean data further; many districts have a per-pupil
expenditure of zero because they are combined with other districts.

However, good progress for this evening.