Improved, readable scatterplot
Turns out cleaning up the data means that D3 automatically re-adjusts
the axes. We had a lot of data with zero, which meant that the MCAS
score axis started at zero, grouping almost all the points way to the left.
Cleaning the data resulted in an axis that started in the 400s.

I also changed margin:left to 100 to make the y-axis readable on the page.

This is ready to be loaded into the D3 template and annotated
as the Week 4 lesson.

Potential extra good stuff: add number of students in district
as a column, and then assign it to the size of the dots.